// libs/ngx-cron-editor/src/enums.d.ts
export declare const Days: {
    SUN: string;
    MON: string;
    TUE: string;
    WED: string;
    THU: string;
    FRI: string;
    SAT: string;
};
export declare const MonthWeeks: {
    '#1': string;
    '#2': string;
    '#3': string;
    '#4': string;
    '#5': string;
    L: string;
};
export declare const Months: {
    1: string;
    2: string;
    3: string;
    4: string;
    5: string;
    6: string;
    7: string;
    8: string;
    9: string;
    10: string;
    11: string;
    12: string;
};